add clarity + TODOs
