+++ src/app/about/page.tsx
@@ -46,7 +46,7 @@ const AboutPage = () => {
       <h3>How to Get Started?</h3>
       <p>
         To interact with PolicyWonk, simply start a chat with{' '}
-        <Link href='/chat/new'>Policy Wonk</Link> and ask it any policy-related
+        <Link href='/'>Policy Wonk</Link> and ask it any policy-related
         question. For example: <code>What are the official holidays?</code>
       </p>
