+++ src/components/chat/rolePortrait.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 
 import Policywonksvg from './policywonksvg';
