+++ src/components/chat/wonkMessage.tsx
@@ -13,23 +13,6 @@ import {
 import Feedback from './feedback';
 import { MemoizedReactMarkdown } from './markdown';
 import { WonkPortrait } from './rolePortrait';
-import { UserPortrait } from './userPortrait';
-
-export const UserMessage = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <div className='row mb-3'>
-      <div className='col-3 col-md-1 mb-2'>
-        <UserPortrait />
-      </div>
-      <div className='col-10 col-md-11'>
-        <p className='chat-name'>
-          <strong>{`You: `}</strong>
-        </p>
-        {children}
-      </div>
-    </div>
-  );
-};
 
 export const WonkMessage = ({
   chatId,
