+++ src/components/layout/sidebar.tsx
@@ -11,7 +11,7 @@ const Sidebar: React.FC = () => {
     <div className='sidebar-container'>
       <header>
         <h1 className='logo'>
-          <Link href='/chat/new'>
+          <Link href='/'>
             Policy Wonk <br />
             <span className='subtitle'>Your UC Policy expert</span>
           </Link>
