+++ src/models/chat.ts
@@ -1,7 +1,5 @@
 import { Message } from 'ai';
 
-export const defaultLlmModel = 'gpt-3.5-turbo';
-
 export type ChatHistory = {
   id: string;
   title: string;
